refctored code into modules
