+++ index.html
@@ -31,7 +31,7 @@
         </div>
         
     </div>
-    <script src="js/feed.js"></script>
-    <script src="js/logout.js"></script>
+    <script type="module" src="src/index.js"></script>
+    <script type="module" src="src/js/logout.js"></script>
 </body>
 </html>
