+++ pages/editPost.html
@@ -32,6 +32,6 @@
             </div>
         </div>
     </div>
-    <script src="../js/editPosts.js"></script> <!-- Link to your JavaScript file for editing posts -->
+    <script type="module" src="../src/js/editPosts.js"></script> <!-- Link to your JavaScript file for editing posts -->
 </body>
 </html>
