+++ pages/login.html
@@ -34,7 +34,7 @@
             </div>
         </div>
     </div>
-    <script src="../js/notifyLogout.js"></script>
-    <script src="../js/login.js"></script>
+    <script type="module" src="../src/js/notifyLogout.js"></script>
+    <script type="module" src="../src/js/login.js"></script>
 </body>
 </html>
