+++ pages/postDetails.html
@@ -14,6 +14,6 @@
         </div>
         <a href="../index.html">Back to posts</a>
     </div>
-    <script src="../js/postDetails.js"></script>
+    <script type="module" src="../src/js/postDetails.js"></script>
 </body>
 </html>
