+++ pages/posts.html
@@ -31,6 +31,6 @@
             </div>
         </div>
     </div>
-    <script src="../js/createPosts.js"></script>
+    <script type="module" src="../src/js/createPosts.js"></script>
 </body>
 </html>
