+++ pages/register.html
@@ -38,6 +38,6 @@
         </div>
     </div>
     
-    <script src="../js/register.js"></script>
+    <script type="module" src="../src/js/register.js"></script>
 </body>
 </html>
