+++ src/js/filterPostModule.js
@@ -1,0 +1,12 @@
+// src/js/filterPostsModule.js
+
+export function filterPosts(allPosts, filterType) {
+    switch (filterType) {
+        case 'withTags':
+            return allPosts.filter(post => post.tags.length > 0);
+        case 'withoutTags':
+            return allPosts.filter(post => post.tags.length === 0);
+        default:
+            return allPosts;
+    }
+}
