+++ src/js/logoutHandler.js
@@ -1,0 +1,6 @@
+// src/js/logoutHandler.js
+
+export function logoutUser() {
+    localStorage.removeItem('userToken');
+    window.location.href = 'pages/login.html';
+}
