+++ src/js/messageDisplay.js
@@ -1,0 +1,13 @@
+// src/js/messageDisplay.js
+
+export function displaySuccessMessage(message) {
+    const successDiv = document.getElementById('successMessage');
+    successDiv.textContent = message;
+    successDiv.classList.remove('d-none');
+}
+
+export function displayErrorMessage(message) {
+    const errorDiv = document.getElementById('errorMessage');
+    errorDiv.textContent = message;
+    errorDiv.classList.remove('d-none');
+}
