+++ src/js/urlParams.js
@@ -1,0 +1,6 @@
+// src/js/urlParams.js
+
+export function getUrlParameter(name) {
+    const urlParams = new URLSearchParams(window.location.search);
+    return urlParams.get(name);
+}
