+++ src/utils/domUtils.js
@@ -1,0 +1,7 @@
+// domUtils.js
+
+export function getElementById(id) {
+    return document.getElementById(id);
+}
+
+// Other DOM utilities as needed...
