+++ src/utils/utils.js
@@ -1,0 +1,11 @@
+// utils.js
+
+export function getFromLocalStorage(key) {
+    return localStorage.getItem(key);
+}
+
+export function setToLocalStorage(key, value) {
+    localStorage.setItem(key, value);
+}
+
+// Other utility functions as needed...
